+++ pkgs/artisanal/lib/plugins.dart
@@ -15,6 +15,7 @@ export 'src/plugins/remote_surface_protocol.dart'
         RemotePluginCursorShape,
         RemotePluginMouseAction,
         RemotePluginMouseButton,
+        RemotePluginNotificationLevel,
         RemotePluginAnchorRect,
         RemotePluginCellAttributes,
         RemotePluginFrameCell,
@@ -29,6 +30,7 @@ export 'src/plugins/remote_surface_protocol.dart'
         RemotePluginClipboardReadRequest,
         RemotePluginClipboardWriteRequest,
         RemotePluginOpenUrlRequest,
+        RemotePluginNotificationRequest,
         RemotePluginKeyInput,
         RemotePluginMouseInput,
         RemotePluginFocusInput,
@@ -36,6 +38,7 @@ export 'src/plugins/remote_surface_protocol.dart'
         RemotePluginClipboardReadResponse,
         RemotePluginClipboardWriteResponse,
         RemotePluginOpenUrlResponse,
+        RemotePluginNotificationResponse,
         RemotePluginProtocolSchemas,
         RemotePluginProtocolValidator,
         RemotePluginProtocolValidationException;
@@ -49,6 +52,8 @@ export 'src/plugins/remote_surface_clipboard_service.dart'
         RemotePluginClipboardHostService;
 export 'src/plugins/remote_surface_url_service.dart'
     show RemotePluginUrlOpener, RemotePluginOpenUrlHostService;
+export 'src/plugins/remote_surface_notification_service.dart'
+    show RemotePluginNotifier, RemotePluginNotificationHostService;
 export 'src/plugins/remote_surface_controller.dart'
     show RemotePluginSurfaceController;
 export 'src/plugins/remote_surface_drawable.dart'